# Alerting: make `/api/prometheus/grafana/api/v1/rules` faster