**What this PR does / why we need it**:

I was testing API perf, and spent some time messing around with `/api/prometheus/grafana/api/v1/rules` to see if there's any low hanging fruit to make it go faster.

By fetching all alert rules in one db query instead of query per group there's about 40% improvement in response time. With ~3000 rules, sqlite 440ms -> 260ms, postgres 1100ms -> 660ms on my machine