@gotjosh 2021-09-27
Can you plesae post the queries we used to do alongside an example query plan and the new query's query plan?