@domasx2 2021-09-27
> Can you plesae post the queries we used to do alongside an example query plan and the new query's query plan?

Did not modify any existing queries or introduce new ones. Refactored API handler to use `GetOrgAlertRules` to get all of the rules at once, instead of sequentially calling `GetRuleGroupAlertRules` for each rule group, essentially a db query per rule. I'll try to get the query plans tomorrow, but I don't think they will be interesting