@gotjosh 2021-09-27
I just want to get a gauge on what changed from the database side, if you turn query logging on, you can see which queries we were doing before vs now, I think that'll give us a good idea of where the improvement is coming from.