@domasx2 2021-09-28
> I just want to get a gauge on what changed from the database side, if you turn query logging on, you can see which queries we were doing before vs now, I think that'll give us a good idea of where the improvement is coming from.

Oki, so, on `main` we do query below for every rule group. And we have group per rule, so in practice for every rule. 3000 rules, 3000 queries.
```sql
SELECT * FROM alert_rule WHERE org_id = $1 and namespace_uid = $2 and rule_group = $
```
Postgres analysis:

```
Index Scan using IDX_alert_rule_org_id_namespace_uid_rule_group on alert_rule as alert_rule (rows=1 loops=1)
Index Cond: ((org_id = 1) AND ((namespace_uid)::text = 'aDUynZI7z'::text) AND ((rule_group)::text = 'Folder-4-94'::text))
```

On this branch we query below query only once to get all rules:

```sql
SELECT * FROM alert_rule WHERE org_id = $1
```

Postges analysis:

```
Seq Scan on alert_rule as alert_rule (rows=3000 loops=1)
Filter: (org_id = 1)
```


I believe the improvement is coming from removing the overhead of doing N quries to db, one for every result item. 
I tried to get a trace, which would better illustrate the point, but it seems only top level http handlers are instrumented...