@gerobinson 2021-09-28
This makes sense to me, and the improvement should be more than 200% for customers who run Grafana and their database on separate hosts. That said I am surprised we are doing `SELECT * FROM alert_rule WHERE org_id = $1` as we will want to avoid doing sequential scans for larger result sets. Perhaps we need to start looking at pagination if we are expecting users to have 1000s of alert rules?