@gerobinson 2021-09-28
I think this change will mean I need to change the database index I am adding in #39476 as we are removing the condition on `namespace_uid`.