@domasx2 2021-09-28
Maybe somewhat contraversially, I think thousands is not a lot and ok to fetch in one go :)

I only have 1 org currently, so seq scan makes sense. I believe if there were many orgs, and the org currently being retrieved occupied a small percentage of the table, an index scan would be chosen. 

Pagination is tricky with the current UX approach, and I think unlikely to have good perf savings.

I think some smaller improvements could easily be had here by removing `GetOrgRuleGroups` and `GetNamespaces` calls above and doing a join with `dashboard` table instead. 

Big improvement could possibly be had by keeping this data in memory and not doing a db call at all - we already have most of it in memory in the alerting engine anyway?